fix: code optimisation performed
