+++ src/components/components.tsx
@@ -1,9 +1,9 @@
-export { HeroSection } from "./hero/heroSection.tsx";
-export { Header } from "./header/header.tsx";
-export { BaseButton } from "./button/baseButton.tsx";
-export { StaffProfiles } from "./staff-profiles/staffProfiles.tsx";
-export { Card } from "./card/card.tsx";
-export { Registration } from "./registration/registration.tsx";
-export { RadioInput } from "./radio-input/radioInput.tsx";
-export { Input } from "./input/input.tsx";
-export { InputFile } from "./input-photo/inputFile.tsx";
+export { HeroSection } from "./sections/hero/heroSection.tsx";
+export { Header } from "./sections/header/header.tsx";
+export { BaseButton } from "./ui/button/baseButton.tsx";
+export { StaffProfiles } from "./sections/staff-profiles/staffProfiles.tsx";
+export { Card } from "./ui/card/card.tsx";
+export { Registration } from "./sections/registration/registration.tsx";
+export { RadioInput } from "./forms/radio-input/radioInput.tsx";
+export { Input } from "./forms/input/input.tsx";
+export { InputFile } from "./forms/input-photo/inputFile.tsx";
