+++ src/components/forms/input-photo/types/type.ts
