+++ src/components/forms/radio-input/radioInput.module.scss
@@ -1,6 +1,3 @@
-.radio {
-}
-
 .label {
 	display: flex;
 	align-items: center;
@@ -14,11 +11,11 @@
 }
 
 .radioCustom {
+	display: flex;
+	position: relative;
 	width: 20px;
 	height: 20px;
 	border: 1px solid $gray-300;
-	display: flex;
-	position: relative;
 	transition: border-color 0.3s ease-in-out;
 }
