+++ src/components/sections/registration/registration.tsx
@@ -4,22 +4,22 @@ import { useForm } from "react-hook-form";
 import { Notify } from "notiflix";
 
 import type { AxiosError } from "axios";
-import type { Position, PositionsResponse } from "../../shared/types/positions-type.ts";
-import type { FormValues } from "../../shared/types/form-registrarion-value.ts";
+import type { Position, PositionsResponse } from "../../../shared/types/positions-type.ts";
+import type { FormValues } from "../../../shared/types/form-registrarion-value.ts";
 import type { RegistrationPropsType } from "./types/type.ts";
-import type { RegisterTokenType } from "../../shared/types/register-token-type.ts";
+import type { RegisterTokenType } from "../../../shared/types/register-token-type.ts";
 import type {
 	SuccessResponse,
 	UserErrorResponseType,
-} from "../../shared/types/user-response-type.ts";
-import type { ErrorResponse } from "../../shared/types/response-error.ts";
+} from "../../../shared/types/user-response-type.ts";
+import type { ErrorResponse } from "../../../shared/types/response-error.ts";
 
-import { EndPointEnum } from "../../shared/enum/end-point-enum.ts";
+import { EndPointEnum } from "../../../shared/enum/end-point-enum.ts";
 
-import { RadioInput, Input, BaseButton, InputFile } from "../components.tsx";
+import { RadioInput, Input, BaseButton, InputFile } from "../../components.tsx";
 
-import { apiRequest } from "../../shared/api";
-import { formFieldsConfig, formFieldsPhotoConfig } from "../../configs/form-fields-config.ts";
+import { apiRequest } from "../../../shared/api";
+import { formFieldsConfig, formFieldsPhotoConfig } from "../../../configs/form-fields-config.ts";
 
 import style from "./registration.module.scss";
 
@@ -35,22 +35,6 @@ const Registration: React.FC<RegistrationPropsType> = ({ setReloadTrigger }) => 
 		mode: "all",
 	});
 
-	useEffect(() => {
-		const fetchPositions = async () => {
-			try {
-				const response = await apiRequest<PositionsResponse>({
-					endpoint: EndPointEnum.POSITIONS,
-				});
-				setPositions(response.positions);
-			} catch (error) {
-				const err = error as AxiosError<ErrorResponse>;
-				Notify.failure(err.message);
-			}
-		};
-
-		void fetchPositions();
-	}, []);
-
 	const onSubmit = async (data: FormValues) => {
 		try {
 			const token = await apiRequest<RegisterTokenType>({
@@ -61,8 +45,8 @@ const Registration: React.FC<RegistrationPropsType> = ({ setReloadTrigger }) => 
 			const formData = new FormData();
 			formData.append("name", data.name);
 			formData.append("email", data.email);
-			formData.append("phone", data.phone.replace(/\D/g, "").trim());
-			formData.append("position_id", String(data.position_id));
+			formData.append("phone", data.phone.replace(/[^\d+]/g, "").trim());
+			formData.append("position_id", data.position_id);
 			formData.append("photo", data.photo[0]);
 
 			const response: SuccessResponse = await apiRequest({
@@ -78,6 +62,7 @@ const Registration: React.FC<RegistrationPropsType> = ({ setReloadTrigger }) => 
 			Notify.success(response.message);
 		} catch (error) {
 			const err = error as AxiosError<UserErrorResponseType>;
+			console.log(err);
 			if (err.response) {
 				Notify.failure(err.response.data.message);
 			} else {
@@ -86,8 +71,26 @@ const Registration: React.FC<RegistrationPropsType> = ({ setReloadTrigger }) => 
 		}
 	};
 
+	const fetchPositions = async () => {
+		try {
+			const response = await apiRequest<PositionsResponse>({
+				endpoint: EndPointEnum.POSITIONS,
+			});
+			setPositions(response.positions);
+		} catch (error) {
+			const err = error as AxiosError<ErrorResponse>;
+			Notify.failure(err.message);
+		}
+	};
+
+	useEffect(() => {
+		void fetchPositions();
+	}, []);
+
 	return (
-		<section className={`container ${style.registration}`}>
+		<section
+			className={`container ${style.registration}`}
+			id="registration">
 			<h2 className={style.title}>Working with POST request</h2>
 			<form
 				className={style.form}
