+++ src/components/sections/registration/types/type.ts
