+++ src/components/sections/staff-profiles/types/type.ts
