+++ src/components/ui/button/baseButton.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import type { Button } from "./types/type.ts";
 
-import styles from "./buton.module.scss";
+import styles from "./button.module.scss";
 
 const BaseButton: React.FC<Button> = ({
 	text,
